requirements.txt
asyncpg
fastapi
grpcio
grpcio-tools
httpx
numpy
psycopg2-binary
pytest
python-dotenv
redis
rich
scikit-learn
strawberry-graphql[fastapi]
tensorflow
uvicorn
websockets
PyYAML
boto3>=1.28.0
botocore>=1.31.0
graphene>=3.3.0
graphql-core>=3.2.3
starlette>=0.27.0
pulumi>=3.75.0
pulumi-aws>=6.0.0
pulumi-kubernetes>=4.0.0
pandas>=2.0.0
scipy>=1.11.0
aiohttp>=3.8.0
aiofiles>=23.0.0
asyncio-mqtt>=0.11.0
prometheus-client>=0.17.0
opentelemetry-api>=1.19.0
opentelemetry-sdk>=1.19.0
cryptography>=41.0.0
pyjwt>=2.8.0
toml>=0.10.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
requests>=2.31.0
click>=8.1.0
tqdm>=4.66.0
schedule>=1.2.0
torch>=2.0.0
transformers>=4.33.0
statsmodels>=0.14.0
# fbprophet>=0.7.0  # Removed due to build complexity
plotly>=5.15.0
dash>=2.12.0
bokeh>=3.2.0
